Made filters module for aperture photometry
